+++ samila/errors.py
@@ -1,2 +1,14 @@
 # -*- coding: utf-8 -*-
 """Samila errors."""
+
+
+class samilaDataError(Exception):
+    """Data error class."""
+
+    pass
+
+
+class samilaGenerateError(Exception):
+    """Generate error class."""
+
+    pass